python-api/pyproject.toml
[project]
name = "knowledge-graph-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sse-starlette>=2.0",
    "google-auth>=2.30",
    "PyJWT>=2.8",
    "cryptography>=42.0",
    "openai>=1.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "httpx>=0.27",
    "python-multipart>=0.0.9",
    "aiofiles>=23.0",
    "tenacity>=8.3",
    "tiktoken>=0.7",
    "rust-core",
    "lancedb>=0.9",
    "prometheus_client>=0.20",
    "openai>=1.40",
    "Pillow>=10.0",
]

[project.optional-dependencies.multimodal]
# Install for image extraction from PDFs
# pdfium2 ships libpdfium as a wheel — no system libs needed on most platforms
dependencies = ["pdfium2>=8.0"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "respx>=0.21",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=80"

[tool.coverage.run]
omit = ["tests/*", "app/core/rust_bridge.py"]
